Struct aws_sdk_batch::client::fluent_builders::CreateComputeEnvironment

pub struct CreateComputeEnvironment<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateComputeEnvironment<C>

pub async fn send(
self
) -> Result<CreateComputeEnvironmentOutput, SdkError<CreateComputeEnvironmentError>> where
C: SmithyConnector,

pub fn compute_environment_name(self, input: impl Into<String>) -> Self

The name for your compute environment. Up to 128 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

pub fn set_compute_environment_name(self, input: Option<String>) -> Self

pub fn type(self, input: CeType) -> Self

The type of the compute environment: MANAGED or UNMANAGED. For more information, see Compute Environments in the AWS Batch User Guide.

pub fn set_type(self, input: Option<CeType>) -> Self

pub fn state(self, input: CeState) -> Self

The state of the compute environment. If the state is ENABLED, then the compute environment accepts jobs from a queue and can scale out automatically based on queues.

If the state is ENABLED, then the AWS Batch scheduler can attempt to place jobs from an associated job queue on the compute resources within the environment. If the compute environment is managed, then it can scale its instances out or in automatically, based on the job queue demand.

If the state is DISABLED, then the AWS Batch scheduler doesn't attempt to place jobs within the environment. Jobs in a STARTING or RUNNING state continue to progress normally. Managed compute environments in the DISABLED state don't scale out. However, they scale in to minvCpus value after instances become idle.

pub fn set_state(self, input: Option<CeState>) -> Self

pub fn compute_resources(self, input: ComputeResource) -> Self

Details about the compute resources managed by the compute environment. This parameter is required for managed compute environments. For more information, see Compute Environments in the AWS Batch User Guide.

pub fn set_compute_resources(self, input: Option<ComputeResource>) -> Self

pub fn service_role(self, input: impl Into<String>) -> Self

The full Amazon Resource Name (ARN) of the IAM role that allows AWS Batch to make calls to other AWS services on your behalf. For more information, see AWS Batch service IAM role in the AWS Batch User Guide.

If your account has already created the AWS Batch service-linked role, that role is used by default for your compute environment unless you specify a role here. If the AWS Batch service-linked role does not exist in your account, and no role is specified here, the service will try to create the AWS Batch service-linked role in your account.

If your specified role has a path other than /, then you must specify either the full role ARN (recommended) or prefix the role name with the path. For example, if a role with the name bar has a path of /foo/ then you would specify /foo/bar as the role name. For more information, see Friendly names and paths in the IAM User Guide.

Depending on how you created your AWS Batch service role, its ARN might contain the service-role path prefix. When you only specify the name of the service role, AWS Batch assumes that your ARN doesn't use the service-role path prefix. Because of this, we recommend that you specify the full ARN of your service role when you create compute environments.

pub fn set_service_role(self, input: Option<String>) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

The tags that you apply to the compute environment to help you categorize and organize your resources. Each tag consists of a key and an optional value. For more information, see Tagging AWS Resources in AWS General Reference.

These tags can be updated or removed using the TagResource and UntagResource API operations. These tags don't propagate to the underlying compute resources.

pub fn set_tags(self, input: Option<HashMap<String, String>>) -> Self