Struct aws_sdk_batch::client::fluent_builders::RegisterJobDefinition

pub struct RegisterJobDefinition<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> RegisterJobDefinition<C>

pub async fn send(
self
) -> Result<RegisterJobDefinitionOutput, SdkError<RegisterJobDefinitionError>> where
C: SmithyConnector,

pub fn job_definition_name(self, input: impl Into<String>) -> Self

The name of the job definition to register. Up to 128 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

pub fn set_job_definition_name(self, input: Option<String>) -> Self

pub fn type(self, input: JobDefinitionType) -> Self

The type of job definition. For more information about multi-node parallel jobs, see Creating a multi-node parallel job definition in the AWS Batch User Guide.

If the job is run on Fargate resources, then multinode isn't supported.

pub fn set_type(self, input: Option<JobDefinitionType>) -> Self

pub fn parameters(self, k: impl Into<String>, v: impl Into<String>) -> Self

Default parameter substitution placeholders to set in the job definition. Parameters are specified as a key-value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition.

pub fn set_parameters(self, input: Option<HashMap<String, String>>) -> Self

pub fn container_properties(self, input: ContainerProperties) -> Self

An object with various properties specific to single-node container-based jobs. If the job definition's type parameter is container, then you must specify either containerProperties or nodeProperties.

If the job runs on Fargate resources, then you must not specify nodeProperties; use only containerProperties.

pub fn set_container_properties(
self,
input: Option<ContainerProperties>
) -> Self

pub fn node_properties(self, input: NodeProperties) -> Self

An object with various properties specific to multi-node parallel jobs. If you specify node properties for a job, it becomes a multi-node parallel job. For more information, see Multi-node Parallel Jobs in the AWS Batch User Guide. If the job definition's type parameter is container, then you must specify either containerProperties or nodeProperties.

If the job runs on Fargate resources, then you must not specify nodeProperties; use containerProperties instead.

pub fn set_node_properties(self, input: Option<NodeProperties>) -> Self

pub fn retry_strategy(self, input: RetryStrategy) -> Self

The retry strategy to use for failed jobs that are submitted with this job definition. Any retry strategy that's specified during a SubmitJob operation overrides the retry strategy defined here. If a job is terminated due to a timeout, it isn't retried.

pub fn set_retry_strategy(self, input: Option<RetryStrategy>) -> Self

pub fn propagate_tags(self, input: bool) -> Self

Specifies whether to propagate the tags from the job or job definition to the corresponding Amazon ECS task. If no value is specified, the tags are not propagated. Tags can only be propagated to the tasks during task creation. For tags with the same name, job tags are given priority over job definitions tags. If the total number of combined tags from the job and job definition is over 50, the job is moved to the FAILED state.

pub fn set_propagate_tags(self, input: Option<bool>) -> Self

pub fn timeout(self, input: JobTimeout) -> Self

The timeout configuration for jobs that are submitted with this job definition, after which AWS Batch terminates your jobs if they have not finished. If a job is terminated due to a timeout, it isn't retried. The minimum value for the timeout is 60 seconds. Any timeout configuration that's specified during a SubmitJob operation overrides the timeout configuration defined here. For more information, see Job Timeouts in the AWS Batch User Guide.

pub fn set_timeout(self, input: Option<JobTimeout>) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

The tags that you apply to the job definition to help you categorize and organize your resources. Each tag consists of a key and an optional value. For more information, see Tagging AWS Resources in AWS Batch User Guide.

pub fn set_tags(self, input: Option<HashMap<String, String>>) -> Self

pub fn platform_capabilities(self, inp: impl Into<PlatformCapability>) -> Self

The platform capabilities required by the job definition. If no value is specified, it defaults to EC2. To run the job on Fargate resources, specify FARGATE.

pub fn set_platform_capabilities(
self,
input: Option<Vec<PlatformCapability>>
) -> Self