Module aws_sdk_batch::error

Modules

client_exception

See ClientException

server_exception

See ServerException

Structs

CancelJobError

ClientException

These errors are usually caused by a client action, such as using an action or resource on behalf of a user that

CreateComputeEnvironmentError

CreateJobQueueError

DeleteComputeEnvironmentError

DeleteJobQueueError

DeregisterJobDefinitionError

DescribeComputeEnvironmentsError

DescribeJobDefinitionsError

DescribeJobQueuesError

DescribeJobsError

ListJobsError

ListTagsForResourceError

RegisterJobDefinitionError

ServerException

These errors are usually caused by a server issue.

SubmitJobError

TagResourceError

TerminateJobError

UntagResourceError

UpdateComputeEnvironmentError

UpdateJobQueueError

Enums

CancelJobErrorKind

CreateComputeEnvironmentErrorKind

CreateJobQueueErrorKind

DeleteComputeEnvironmentErrorKind

DeleteJobQueueErrorKind

DeregisterJobDefinitionErrorKind

DescribeComputeEnvironmentsErrorKind

DescribeJobDefinitionsErrorKind

DescribeJobQueuesErrorKind

DescribeJobsErrorKind

ListJobsErrorKind

ListTagsForResourceErrorKind

RegisterJobDefinitionErrorKind

SubmitJobErrorKind

TagResourceErrorKind

TerminateJobErrorKind

UntagResourceErrorKind

UpdateComputeEnvironmentErrorKind

UpdateJobQueueErrorKind