Module aws_sdk_batch::model

Modules

array_properties

See ArrayProperties

array_properties_detail

See ArrayPropertiesDetail

array_properties_summary

See ArrayPropertiesSummary

attempt_container_detail

See AttemptContainerDetail

attempt_detail

See AttemptDetail

compute_environment_detail

See ComputeEnvironmentDetail

compute_environment_order

See ComputeEnvironmentOrder

compute_resource

See ComputeResource

compute_resource_update

See ComputeResourceUpdate

container_detail

See ContainerDetail

container_overrides

See ContainerOverrides

container_properties

See ContainerProperties

container_summary

See ContainerSummary

device

See Device

ec2_configuration

See Ec2Configuration

efs_authorization_config

See EfsAuthorizationConfig

efs_volume_configuration

See EfsVolumeConfiguration

evaluate_on_exit

See EvaluateOnExit

fargate_platform_configuration

See FargatePlatformConfiguration

host

See Host

job_definition

See JobDefinition

job_dependency

See JobDependency

job_detail

See JobDetail

job_queue_detail

See JobQueueDetail

job_summary

See JobSummary

job_timeout

See JobTimeout

key_value_pair

See KeyValuePair

launch_template_specification

See LaunchTemplateSpecification

linux_parameters

See LinuxParameters

log_configuration

See LogConfiguration

mount_point

See MountPoint

network_configuration

See NetworkConfiguration

network_interface

See NetworkInterface

node_details

See NodeDetails

node_overrides

See NodeOverrides

node_properties

See NodeProperties

node_properties_summary

See NodePropertiesSummary

node_property_override

See NodePropertyOverride

node_range_property

See NodeRangeProperty

resource_requirement

See ResourceRequirement

retry_strategy

See RetryStrategy

secret

See Secret

tmpfs

See Tmpfs

ulimit

See Ulimit

volume

See Volume

Structs

ArrayProperties

An object representing an AWS Batch array job.

ArrayPropertiesDetail

An object representing the array properties of a job.

ArrayPropertiesSummary

An object representing the array properties of a job.

AttemptContainerDetail

An object representing the details of a container that's part of a job attempt.

AttemptDetail

An object representing a job attempt.

ComputeEnvironmentDetail

An object representing an AWS Batch compute environment.

ComputeEnvironmentOrder

The order in which compute environments are tried for job placement within a queue. Compute environments are

ComputeResource

An object representing an AWS Batch compute resource. For more information, see Compute Environments in the

ComputeResourceUpdate

An object representing the attributes of a compute environment that can be updated. For more information, see

ContainerDetail

An object representing the details of a container that's part of a job.

ContainerOverrides

The overrides that should be sent to a container.

ContainerProperties

Container properties are used in job definitions to describe the container that's launched as part of a

ContainerSummary

An object representing summary details of a container within a job.

Device

An object representing a container instance host device.

Ec2Configuration

Provides information used to select Amazon Machine Images (AMIs) for instances in the compute environment. If

EfsAuthorizationConfig

The authorization configuration details for the Amazon EFS file system.

EfsVolumeConfiguration

This parameter is specified when you are using an Amazon Elastic File System file system for task storage. For more information,

EvaluateOnExit

Specifies a set of conditions to be met, and an action to take (RETRY or EXIT) if all

FargatePlatformConfiguration

The platform configuration for jobs running on Fargate resources. For jobs that run on EC2 resources, you

Host

Determine whether your data volume persists on the host container instance and where it is stored. If this

JobDefinition

An object representing an AWS Batch job definition.

JobDependency

An object representing an AWS Batch job dependency.

JobDetail

An object representing an AWS Batch job.

JobQueueDetail

An object representing the details of an AWS Batch job queue.

JobSummary

An object representing summary details of a job.

JobTimeout

An object representing a job timeout configuration.

KeyValuePair

A key-value pair object.

LaunchTemplateSpecification

An object representing a launch template associated with a compute resource. You must specify either the launch

LinuxParameters

Linux-specific modifications that are applied to the container, such as details for device mappings.

LogConfiguration

Log configuration options to send to a custom log driver for the container.

MountPoint

Details on a Docker volume mount point that's used in a job's container properties. This parameter maps to

NetworkConfiguration

The network configuration for jobs running on Fargate resources. Jobs running on EC2 resources must not

NetworkInterface

An object representing the elastic network interface for a multi-node parallel job node.

NodeDetails

An object representing the details of a multi-node parallel job node.

NodeOverrides

Object representing any node overrides to a job definition that's used in a SubmitJob API

NodeProperties

An object representing the node properties of a multi-node parallel job.

NodePropertiesSummary

An object representing the properties of a node that's associated with a multi-node parallel job.

NodePropertyOverride

Object representing any node overrides to a job definition that's used in a SubmitJob API

NodeRangeProperty

An object representing the properties of the node range for a multi-node parallel job.

ResourceRequirement

The type and amount of a resource to assign to a container. The supported resources include GPU,

RetryStrategy

The retry strategy associated with a job. For more information, see Automated job retries in the AWS Batch User Guide.

Secret

An object representing the secret to expose to your container. Secrets can be exposed to a container in the

Tmpfs

The container path, mount options, and size of the tmpfs mount.

Ulimit

The ulimit settings to pass to the container.

Volume

A data volume used in a job's container properties.

Enums

ArrayJobDependency

AssignPublicIp

CeState

CeStatus

CeType

CrAllocationStrategy

CrType

DeviceCgroupPermission

EfsAuthorizationConfigIam

EfsTransitEncryption

JobDefinitionType

JobStatus

JqState

JqStatus

LogDriver

PlatformCapability

ResourceType

RetryAction