Struct aws_sdk_batch::model::container_detail::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ContainerDetail

Implementations

impl Builder

pub fn image(self, input: impl Into<String>) -> Self

The image used to start the container.

pub fn set_image(self, input: Option<String>) -> Self

pub fn vcpus(self, input: i32) -> Self

The number of vCPUs reserved for the container. For jobs that run on EC2 resources, you can specify the vCPU requirement for the job using resourceRequirements, but you can't specify the vCPU requirements in both the vcpus and resourceRequirement object. This parameter maps to CpuShares in the Create a container section of the Docker Remote API and the --cpu-shares option to docker run. Each vCPU is equivalent to 1,024 CPU shares. You must specify at least one vCPU. This is required but can be specified in several places. It must be specified for each node at least once.

This parameter isn't applicable to jobs that run on Fargate resources. For jobs that run on Fargate resources, you must specify the vCPU requirement for the job using resourceRequirements.

pub fn set_vcpus(self, input: Option<i32>) -> Self

pub fn memory(self, input: i32) -> Self

For jobs run on EC2 resources that didn't specify memory requirements using ResourceRequirement, the number of MiB of memory reserved for the job. For other jobs, including all run on Fargate resources, see resourceRequirements.

pub fn set_memory(self, input: Option<i32>) -> Self

pub fn command(self, input: impl Into<String>) -> Self

pub fn set_command(self, input: Option<Vec<String>>) -> Self

pub fn job_role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) associated with the job upon execution.

pub fn set_job_role_arn(self, input: Option<String>) -> Self

pub fn execution_role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the execution role that AWS Batch can assume. For more information, see AWS Batch execution IAM role in the AWS Batch User Guide.

pub fn set_execution_role_arn(self, input: Option<String>) -> Self

pub fn volumes(self, input: impl Into<Volume>) -> Self

pub fn set_volumes(self, input: Option<Vec<Volume>>) -> Self

pub fn environment(self, input: impl Into<KeyValuePair>) -> Self

pub fn set_environment(self, input: Option<Vec<KeyValuePair>>) -> Self

pub fn mount_points(self, input: impl Into<MountPoint>) -> Self

pub fn set_mount_points(self, input: Option<Vec<MountPoint>>) -> Self

pub fn readonly_root_filesystem(self, input: bool) -> Self

When this parameter is true, the container is given read-only access to its root file system. This parameter maps to ReadonlyRootfs in the Create a container section of the Docker Remote API and the --read-only option to docker run .

pub fn set_readonly_root_filesystem(self, input: Option<bool>) -> Self

pub fn ulimits(self, input: impl Into<Ulimit>) -> Self

pub fn set_ulimits(self, input: Option<Vec<Ulimit>>) -> Self

pub fn privileged(self, input: bool) -> Self

When this parameter is true, the container is given elevated permissions on the host container instance (similar to the root user). The default value is false.

This parameter isn't applicable to jobs running on Fargate resources and shouldn't be provided, or specified as false.

pub fn set_privileged(self, input: Option<bool>) -> Self

pub fn user(self, input: impl Into<String>) -> Self

The user name to use inside the container. This parameter maps to User in the Create a container section of the Docker Remote API and the --user option to docker run.

pub fn set_user(self, input: Option<String>) -> Self

pub fn exit_code(self, input: i32) -> Self

The exit code to return upon completion.

pub fn set_exit_code(self, input: Option<i32>) -> Self

pub fn reason(self, input: impl Into<String>) -> Self

A short (255 max characters) human-readable string to provide additional details about a running or stopped container.

pub fn set_reason(self, input: Option<String>) -> Self

pub fn container_instance_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the container instance that the container is running on.

pub fn set_container_instance_arn(self, input: Option<String>) -> Self

pub fn task_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Amazon ECS task that's associated with the container job. Each container attempt receives a task ARN when they reach the STARTING status.

pub fn set_task_arn(self, input: Option<String>) -> Self

pub fn log_stream_name(self, input: impl Into<String>) -> Self

The name of the CloudWatch Logs log stream associated with the container. The log group for AWS Batch jobs is /aws/batch/job. Each container attempt receives a log stream name when they reach the RUNNING status.

pub fn set_log_stream_name(self, input: Option<String>) -> Self

pub fn instance_type(self, input: impl Into<String>) -> Self

The instance type of the underlying host infrastructure of a multi-node parallel job.

This parameter isn't applicable to jobs running on Fargate resources.

pub fn set_instance_type(self, input: Option<String>) -> Self

pub fn network_interfaces(self, input: impl Into<NetworkInterface>) -> Self

pub fn set_network_interfaces(
self,
input: Option<Vec<NetworkInterface>>
) -> Self

pub fn resource_requirements(
self,
input: impl Into<ResourceRequirement>
) -> Self

pub fn set_resource_requirements(
self,
input: Option<Vec<ResourceRequirement>>
) -> Self

pub fn linux_parameters(self, input: LinuxParameters) -> Self

Linux-specific modifications that are applied to the container, such as details for device mappings.

pub fn set_linux_parameters(self, input: Option<LinuxParameters>) -> Self

pub fn log_configuration(self, input: LogConfiguration) -> Self

The log configuration specification for the container.

This parameter maps to LogConfig in the Create a container section of the Docker Remote API and the --log-driver option to docker run. By default, containers use the same logging driver that the Docker daemon uses. However, the container might use a different logging driver than the Docker daemon by specifying a log driver with this parameter in the container definition. To use a different logging driver for a container, the log system must be configured properly on the container instance. Or, alternatively, it must be configured on a different log server for remote logging options. For more information on the options for different supported log drivers, see Configure logging drivers in the Docker documentation.

AWS Batch currently supports a subset of the logging drivers available to the Docker daemon (shown in the LogConfiguration data type). Additional log drivers might be available in future releases of the Amazon ECS container agent.

This parameter requires version 1.18 of the Docker Remote API or greater on your container instance. To check the Docker Remote API version on your container instance, log into your container instance and run the following command: sudo docker version | grep "Server API version"

The Amazon ECS container agent running on a container instance must register the logging drivers available on that instance with the ECS_AVAILABLE_LOGGING_DRIVERS environment variable before containers placed on that instance can use these log configuration options. For more information, see Amazon ECS Container Agent Configuration in the Amazon Elastic Container Service Developer Guide.

pub fn set_log_configuration(self, input: Option<LogConfiguration>) -> Self

pub fn secrets(self, input: impl Into<Secret>) -> Self

pub fn set_secrets(self, input: Option<Vec<Secret>>) -> Self

pub fn network_configuration(self, input: NetworkConfiguration) -> Self

The network configuration for jobs running on Fargate resources. Jobs running on EC2 resources must not specify this parameter.

pub fn set_network_configuration(
self,
input: Option<NetworkConfiguration>
) -> Self

pub fn fargate_platform_configuration(
self,
input: FargatePlatformConfiguration
) -> Self

The platform configuration for jobs running on Fargate resources. Jobs running on EC2 resources must not specify this parameter.

pub fn set_fargate_platform_configuration(
self,
input: Option<FargatePlatformConfiguration>
) -> Self

pub fn build(self) -> ContainerDetail

Consumes the builder and constructs a ContainerDetail