Struct aws_sdk_batch::model::ContainerDetail

#[non_exhaustive]
pub struct ContainerDetail {

    pub image: Option<String>,
    pub vcpus: i32,
    pub memory: i32,
    pub command: Option<Vec<String>>,
    pub job_role_arn: Option<String>,
    pub execution_role_arn: Option<String>,
    pub volumes: Option<Vec<Volume>>,
    pub environment: Option<Vec<KeyValuePair>>,
    pub mount_points: Option<Vec<MountPoint>>,
    pub readonly_root_filesystem: bool,
    pub ulimits: Option<Vec<Ulimit>>,
    pub privileged: bool,
    pub user: Option<String>,
    pub exit_code: i32,
    pub reason: Option<String>,
    pub container_instance_arn: Option<String>,
    pub task_arn: Option<String>,
    pub log_stream_name: Option<String>,
    pub instance_type: Option<String>,
    pub network_interfaces: Option<Vec<NetworkInterface>>,
    pub resource_requirements: Option<Vec<ResourceRequirement>>,
    pub linux_parameters: Option<LinuxParameters>,
    pub log_configuration: Option<LogConfiguration>,
    pub secrets: Option<Vec<Secret>>,
    pub network_configuration: Option<NetworkConfiguration>,
    pub fargate_platform_configuration: Option<FargatePlatformConfiguration>,

}

An object representing the details of a container that's part of a job.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

image: Option<String>

The image used to start the container.

vcpus: i32

The number of vCPUs reserved for the container. For jobs that run on EC2 resources, you can specify the vCPU requirement for the job using resourceRequirements, but you can't specify the vCPU requirements in both the vcpus and resourceRequirement object. This parameter maps to CpuShares in the Create a container section of the Docker Remote API and the --cpu-shares option to docker run. Each vCPU is equivalent to 1,024 CPU shares. You must specify at least one vCPU. This is required but can be specified in several places. It must be specified for each node at least once.

This parameter isn't applicable to jobs that run on Fargate resources. For jobs that run on Fargate resources, you must specify the vCPU requirement for the job using resourceRequirements.

memory: i32

For jobs run on EC2 resources that didn't specify memory requirements using ResourceRequirement, the number of MiB of memory reserved for the job. For other jobs, including all run on Fargate resources, see resourceRequirements.

command: Option<Vec<String>>

The command that's passed to the container.

job_role_arn: Option<String>

The Amazon Resource Name (ARN) associated with the job upon execution.

execution_role_arn: Option<String>

The Amazon Resource Name (ARN) of the execution role that AWS Batch can assume. For more information, see AWS Batch execution IAM role in the AWS Batch User Guide.

volumes: Option<Vec<Volume>>

A list of volumes associated with the job.

environment: Option<Vec<KeyValuePair>>

The environment variables to pass to a container.

Environment variables must not start with AWS_BATCH; this naming convention is reserved for variables that are set by the AWS Batch service.

mount_points: Option<Vec<MountPoint>>

The mount points for data volumes in your container.

readonly_root_filesystem: bool

When this parameter is true, the container is given read-only access to its root file system. This parameter maps to ReadonlyRootfs in the Create a container section of the Docker Remote API and the --read-only option to docker run .

ulimits: Option<Vec<Ulimit>>

A list of ulimit values to set in the container. This parameter maps to Ulimits in the Create a container section of the Docker Remote API and the --ulimit option to docker run.

This parameter isn't applicable to jobs running on Fargate resources.

privileged: bool

When this parameter is true, the container is given elevated permissions on the host container instance (similar to the root user). The default value is false.

This parameter isn't applicable to jobs running on Fargate resources and shouldn't be provided, or specified as false.

user: Option<String>

The user name to use inside the container. This parameter maps to User in the Create a container section of the Docker Remote API and the --user option to docker run.

exit_code: i32

The exit code to return upon completion.

reason: Option<String>

A short (255 max characters) human-readable string to provide additional details about a running or stopped container.

container_instance_arn: Option<String>

The Amazon Resource Name (ARN) of the container instance that the container is running on.

task_arn: Option<String>

The Amazon Resource Name (ARN) of the Amazon ECS task that's associated with the container job. Each container attempt receives a task ARN when they reach the STARTING status.

log_stream_name: Option<String>

The name of the CloudWatch Logs log stream associated with the container. The log group for AWS Batch jobs is /aws/batch/job. Each container attempt receives a log stream name when they reach the RUNNING status.

instance_type: Option<String>

The instance type of the underlying host infrastructure of a multi-node parallel job.

This parameter isn't applicable to jobs running on Fargate resources.

network_interfaces: Option<Vec<NetworkInterface>>

The network interfaces associated with the job.

resource_requirements: Option<Vec<ResourceRequirement>>

The type and amount of resources to assign to a container. The supported resources include GPU, MEMORY, and VCPU.

linux_parameters: Option<LinuxParameters>

Linux-specific modifications that are applied to the container, such as details for device mappings.

log_configuration: Option<LogConfiguration>

The log configuration specification for the container.

This parameter maps to LogConfig in the Create a container section of the Docker Remote API and the --log-driver option to docker run. By default, containers use the same logging driver that the Docker daemon uses. However, the container might use a different logging driver than the Docker daemon by specifying a log driver with this parameter in the container definition. To use a different logging driver for a container, the log system must be configured properly on the container instance. Or, alternatively, it must be configured on a different log server for remote logging options. For more information on the options for different supported log drivers, see Configure logging drivers in the Docker documentation.

AWS Batch currently supports a subset of the logging drivers available to the Docker daemon (shown in the LogConfiguration data type). Additional log drivers might be available in future releases of the Amazon ECS container agent.

This parameter requires version 1.18 of the Docker Remote API or greater on your container instance. To check the Docker Remote API version on your container instance, log into your container instance and run the following command: sudo docker version | grep "Server API version"

The Amazon ECS container agent running on a container instance must register the logging drivers available on that instance with the ECS_AVAILABLE_LOGGING_DRIVERS environment variable before containers placed on that instance can use these log configuration options. For more information, see Amazon ECS Container Agent Configuration in the Amazon Elastic Container Service Developer Guide.

secrets: Option<Vec<Secret>>

The secrets to pass to the container. For more information, see Specifying sensitive data in the AWS Batch User Guide.

network_configuration: Option<NetworkConfiguration>

The network configuration for jobs running on Fargate resources. Jobs running on EC2 resources must not specify this parameter.

fargate_platform_configuration: Option<FargatePlatformConfiguration>

The platform configuration for jobs running on Fargate resources. Jobs running on EC2 resources must not specify this parameter.

Implementations

impl ContainerDetail

pub fn builder() -> Builder

Creates a new builder-style object to manufacture ContainerDetail