Struct aws_sdk_batch::model::ContainerProperties

#[non_exhaustive]
pub struct ContainerProperties {

    pub image: Option<String>,
    pub vcpus: i32,
    pub memory: i32,
    pub command: Option<Vec<String>>,
    pub job_role_arn: Option<String>,
    pub execution_role_arn: Option<String>,
    pub volumes: Option<Vec<Volume>>,
    pub environment: Option<Vec<KeyValuePair>>,
    pub mount_points: Option<Vec<MountPoint>>,
    pub readonly_root_filesystem: bool,
    pub privileged: bool,
    pub ulimits: Option<Vec<Ulimit>>,
    pub user: Option<String>,
    pub instance_type: Option<String>,
    pub resource_requirements: Option<Vec<ResourceRequirement>>,
    pub linux_parameters: Option<LinuxParameters>,
    pub log_configuration: Option<LogConfiguration>,
    pub secrets: Option<Vec<Secret>>,
    pub network_configuration: Option<NetworkConfiguration>,
    pub fargate_platform_configuration: Option<FargatePlatformConfiguration>,

}

Container properties are used in job definitions to describe the container that's launched as part of a job.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

image: Option<String>

The image used to start a container. This string is passed directly to the Docker daemon. Images in the Docker Hub registry are available by default. Other repositories are specified with repository-url/image:tag . Up to 255 letters (uppercase and lowercase), numbers, hyphens, underscores, colons, periods, forward slashes, and number signs are allowed. This parameter maps to Image in the Create a container section of the Docker Remote API and the IMAGE parameter of docker run.

Docker image architecture must match the processor architecture of the compute resources that they're scheduled on. For example, ARM-based Docker images can only run on ARM-based compute resources.

Images in Amazon ECR repositories use the full registry and repository URI (for example, 012345678910.dkr.ecr..amazonaws.com/).

Images in official repositories on Docker Hub use a single name (for example, ubuntu or mongo).

Images in other repositories on Docker Hub are qualified with an organization name (for example, amazon/amazon-ecs-agent).

Images in other online repositories are qualified further by a domain name (for example, quay.io/assemblyline/ubuntu).

vcpus: i32

The number of vCPUs reserved for the job. Each vCPU is equivalent to 1,024 CPU shares. This parameter maps to CpuShares in the Create a container section of the Docker Remote API and the --cpu-shares option to docker run. The number of vCPUs must be specified but can be be specified in several places. You must specify it at least once for each node.

This parameter is supported on EC2 resources but isn't supported for jobs that run on Fargate resources. For these resources, use resourceRequirement instead. You can use this parameter or resourceRequirements structure but not both.

This parameter isn't applicable to jobs running on Fargate resources and shouldn't be provided. For jobs that run on Fargate resources, you must specify the vCPU requirement for the job using resourceRequirements.

memory: i32

This parameter indicates the memory hard limit (in MiB) for a container. If your container attempts to exceed the specified number, it is terminated. You must specify at least 4 MiB of memory for a job using this parameter. The memory hard limit can be specified in several places. It must be specified for each node at least once.

This parameter maps to Memory in the Create a container section of the Docker Remote API and the --memory option to docker run.

This parameter is supported on EC2 resources but isn't supported on Fargate resources. For Fargate resources, you should specify the memory requirement using resourceRequirement. You can do this for EC2 resources.

If you're trying to maximize your resource utilization by providing your jobs as much memory as possible for a particular instance type, see Memory Management in the AWS Batch User Guide.

command: Option<Vec<String>>

The command that's passed to the container. This parameter maps to Cmd in the Create a container section of the Docker Remote API and the COMMAND parameter to docker run. For more information, see https://docs.docker.com/engine/reference/builder/#cmd.

job_role_arn: Option<String>

The Amazon Resource Name (ARN) of the IAM role that the container can assume for AWS permissions. For more information, see IAM Roles for Tasks in the Amazon Elastic Container Service Developer Guide.

execution_role_arn: Option<String>

The Amazon Resource Name (ARN) of the execution role that AWS Batch can assume. For jobs that run on Fargate resources, you must provide an execution role. For more information, see AWS Batch execution IAM role in the AWS Batch User Guide.

volumes: Option<Vec<Volume>>

A list of data volumes used in a job.

environment: Option<Vec<KeyValuePair>>

The environment variables to pass to a container. This parameter maps to Env in the Create a container section of the Docker Remote API and the --env option to docker run.

We don't recommend using plaintext environment variables for sensitive information, such as credential data.

Environment variables must not start with AWS_BATCH; this naming convention is reserved for variables that are set by the AWS Batch service.

mount_points: Option<Vec<MountPoint>>

The mount points for data volumes in your container. This parameter maps to Volumes in the Create a container section of the Docker Remote API and the --volume option to docker run.

readonly_root_filesystem: bool

When this parameter is true, the container is given read-only access to its root file system. This parameter maps to ReadonlyRootfs in the Create a container section of the Docker Remote API and the --read-only option to docker run.

privileged: bool

When this parameter is true, the container is given elevated permissions on the host container instance (similar to the root user). This parameter maps to Privileged in the Create a container section of the Docker Remote API and the --privileged option to docker run. The default value is false.

This parameter isn't applicable to jobs running on Fargate resources and shouldn't be provided, or specified as false.

ulimits: Option<Vec<Ulimit>>

A list of ulimits to set in the container. This parameter maps to Ulimits in the Create a container section of the Docker Remote API and the --ulimit option to docker run.

This parameter isn't applicable to jobs running on Fargate resources and shouldn't be provided.

user: Option<String>

The user name to use inside the container. This parameter maps to User in the Create a container section of the Docker Remote API and the --user option to docker run.

instance_type: Option<String>

The instance type to use for a multi-node parallel job. All node groups in a multi-node parallel job must use the same instance type.

This parameter isn't applicable to single-node container jobs or for jobs that run on Fargate resources and shouldn't be provided.

resource_requirements: Option<Vec<ResourceRequirement>>

The type and amount of resources to assign to a container. The supported resources include GPU, MEMORY, and VCPU.

linux_parameters: Option<LinuxParameters>

Linux-specific modifications that are applied to the container, such as details for device mappings.

log_configuration: Option<LogConfiguration>

The log configuration specification for the container.

This parameter maps to LogConfig in the Create a container section of the Docker Remote API and the --log-driver option to docker run. By default, containers use the same logging driver that the Docker daemon uses. However the container might use a different logging driver than the Docker daemon by specifying a log driver with this parameter in the container definition. To use a different logging driver for a container, the log system must be configured properly on the container instance (or on a different log server for remote logging options). For more information on the options for different supported log drivers, see Configure logging drivers in the Docker documentation.

AWS Batch currently supports a subset of the logging drivers available to the Docker daemon (shown in the LogConfiguration data type).

This parameter requires version 1.18 of the Docker Remote API or greater on your container instance. To check the Docker Remote API version on your container instance, log into your container instance and run the following command: sudo docker version | grep "Server API version"

The Amazon ECS container agent running on a container instance must register the logging drivers available on that instance with the ECS_AVAILABLE_LOGGING_DRIVERS environment variable before containers placed on that instance can use these log configuration options. For more information, see Amazon ECS Container Agent Configuration in the Amazon Elastic Container Service Developer Guide.

secrets: Option<Vec<Secret>>

The secrets for the container. For more information, see Specifying sensitive data in the AWS Batch User Guide.

network_configuration: Option<NetworkConfiguration>

The network configuration for jobs running on Fargate resources. Jobs running on EC2 resources must not specify this parameter.

fargate_platform_configuration: Option<FargatePlatformConfiguration>

The platform configuration for jobs running on Fargate resources. Jobs running on EC2 resources must not specify this parameter.

Implementations

impl ContainerProperties

pub fn builder() -> Builder

Creates a new builder-style object to manufacture ContainerProperties