Struct aws_sdk_batch::client::fluent_builders::SubmitJob

pub struct SubmitJob<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> SubmitJob<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(self) -> Result<SubmitJobOutput, SdkError<SubmitJobError>> where
R::Policy: SmithyRetryPolicy<SubmitJobInputOperationOutputAlias, SubmitJobOutput, SubmitJobError, SubmitJobInputOperationRetryAlias>,

pub fn job_name(self, inp: impl Into<String>) -> Self

The name of the job. The first character must be alphanumeric, and up to 128 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

pub fn set_job_name(self, input: Option<String>) -> Self

pub fn job_queue(self, inp: impl Into<String>) -> Self

The job queue where the job is submitted. You can specify either the name or the Amazon Resource Name (ARN) of the queue.

pub fn set_job_queue(self, input: Option<String>) -> Self

pub fn array_properties(self, inp: ArrayProperties) -> Self

The array properties for the submitted job, such as the size of the array. The array size can be between 2 and 10,000. If you specify array properties for a job, it becomes an array job. For more information, see Array Jobs in the Batch User Guide.

pub fn set_array_properties(self, input: Option<ArrayProperties>) -> Self

pub fn depends_on(self, inp: impl Into<JobDependency>) -> Self

Appends an item to dependsOn.

To override the contents of this collection use set_depends_on.

A list of dependencies for the job. A job can depend upon a maximum of 20 jobs. You can specify a SEQUENTIAL type dependency without specifying a job ID for array jobs so that each child array job completes sequentially, starting at index 0. You can also specify an N_TO_N type dependency with a job ID for array jobs. In that case, each index child of this job must wait for the corresponding index child of each dependency to complete before it can begin.

pub fn set_depends_on(self, input: Option<Vec<JobDependency>>) -> Self

pub fn job_definition(self, inp: impl Into<String>) -> Self

The job definition used by this job. This value can be one of name, name:revision, or the Amazon Resource Name (ARN) for the job definition. If name is specified without a revision then the latest active revision is used.

pub fn set_job_definition(self, input: Option<String>) -> Self

pub fn parameters(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to parameters.

To override the contents of this collection use set_parameters.

Additional parameters passed to the job that replace parameter substitution placeholders that are set in the job definition. Parameters are specified as a key and value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition.

pub fn set_parameters(self, input: Option<HashMap<String, String>>) -> Self

pub fn container_overrides(self, inp: ContainerOverrides) -> Self

A list of container overrides in the JSON format that specify the name of a container in the specified job definition and the overrides it should receive. You can override the default command for a container, which is specified in the job definition or the Docker image, with a command override. You can also override existing environment variables on a container or add new environment variables to it with an environment override.

pub fn set_container_overrides(self, input: Option<ContainerOverrides>) -> Self

pub fn node_overrides(self, inp: NodeOverrides) -> Self

A list of node overrides in JSON format that specify the node range to target and the container overrides for that node range.

This parameter isn't applicable to jobs that are running on Fargate resources; use containerOverrides instead.

pub fn set_node_overrides(self, input: Option<NodeOverrides>) -> Self

pub fn retry_strategy(self, inp: RetryStrategy) -> Self

The retry strategy to use for failed jobs from this SubmitJob operation. When a retry strategy is specified here, it overrides the retry strategy defined in the job definition.

pub fn set_retry_strategy(self, input: Option<RetryStrategy>) -> Self

pub fn propagate_tags(self, inp: bool) -> Self

Specifies whether to propagate the tags from the job or job definition to the corresponding Amazon ECS task. If no value is specified, the tags aren't propagated. Tags can only be propagated to the tasks during task creation. For tags with the same name, job tags are given priority over job definitions tags. If the total number of combined tags from the job and job definition is over 50, the job is moved to the FAILED state. When specified, this overrides the tag propagation setting in the job definition.

pub fn set_propagate_tags(self, input: Option<bool>) -> Self

pub fn timeout(self, inp: JobTimeout) -> Self

The timeout configuration for this SubmitJob operation. You can specify a timeout duration after which Batch terminates your jobs if they haven't finished. If a job is terminated due to a timeout, it isn't retried. The minimum value for the timeout is 60 seconds. This configuration overrides any timeout configuration specified in the job definition. For array jobs, child jobs have the same timeout configuration as the parent job. For more information, see Job Timeouts in the Amazon Elastic Container Service Developer Guide.

pub fn set_timeout(self, input: Option<JobTimeout>) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to tags.

To override the contents of this collection use set_tags.

The tags that you apply to the job request to help you categorize and organize your resources. Each tag consists of a key and an optional value. For more information, see Tagging Amazon Web Services Resources in Amazon Web Services General Reference.

pub fn set_tags(self, input: Option<HashMap<String, String>>) -> Self